rename indicatorShader.js to indicatorShaders.js
